requirements.txt
azure-common==1.1.27
azure-core==1.13.0
azure-identity==1.5.0
azure-mgmt-compute==20.0.0
azure-mgmt-core==1.2.2
azure-mgmt-network==18.0.0
azure-mgmt-resource==16.1.0
bcrypt==3.2.0
beautifulsoup4==4.7.1
boto
boto3==1.17.61
botocore==1.20.61
bs4==0.0.1
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
cmake==3.18.4.post1
colorama==0.3.7
colorclass==2.2.0


contextvars==2.4
cryptography==3.3.1
cycler==0.10.0
Cython==0.29.21
dask==2020.12.0
dataclasses==0.6
DateTime==4.3
decorator==4.4.2
decord==0.4.2
dill==0.3.3
distributed==2020.12.0
distro==1.5.0

docutils==0.16
enum34==1.1.6
filelock==3.0.12
Flask==1.1.2
flask-limiter==1.5
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.0
flatbuffers==1.12
future==0.18.2
gast==0.2.2
gcloud==0.18.3
gevent==1.4.0
gluoncv==0.7.0
google-api-core==1.21.0
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-storage==1.24.1
google-cloud-vision==0.41.0
google-pasta==0.2.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
graphviz==0.8.4
greenlet==0.4.16
greenstalk==0.5.2
grpcio==1.30.0
gunicorn==19.9.0
h5py==2.10.0
HeapDict==1.0.1
httplib2==0.18.1
idna==2.10
imageio==2.9.0
immutables==0.14
importlib-metadata==2.0.0
iniconfig==1.1.1
isodate==0.6.0
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1

launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
libsvm==3.23.0.4
linode-cli==2.2.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mpmath==1.2.1
msal==1.11.0
msal-extensions==0.3.0
msgpack==0.6.2
msgpack-python==0.5.6
msrest==0.6.21
mxnet==1.6.0
netifaces==0.10.4
networkx==2.5
nltk==3.5
numpy==1.19.2
oauth==1.0.1
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.45.1
opencv-contrib-python==4.2.0.34
opencv-python==4.1.2.30
opt-einsum==3.2.1
packaging==20.8
pandas==1.1.5
paramiko==2.7.2
pexpect==4.2.1
pickle-mixin==1.0.2
pika==0.13.0
Pillow==7.0.0
pluggy==0.13.1
portalocker==1.7.1
promise==2.3
protobuf==3.14.0
psutil==5.7.3
psycopg2-binary==2.8.6
py==1.10.0
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8

pycparser==2.20
pycrypto==2.6.1

Pygments==2.2.0

PyJWT==2.0.1
pylint==2.3.1
pymacaroons==0.13.0
pymongo==3.11.2
PyMySQL==0.9.3
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.0
pytest==6.2.0
python-dateutil==2.8.1
python-debian==0.1.32
python-dotenv==0.10.3
pytz==2020.1
PyWavelets
pyxdg
PyYAML==5.3.1
rank-bm25==0.2.1
redis==3.5.3
regex==2020.11.13
reportlab==3.4.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
roman==2.0.0
rsa==4.6
s3transfer==0.4.0
sacremoses==0.0.43
sacrud==0.3.7.post0
scikit-build==0.11.1
scikit-image==0.15.0
scikit-learn==0.23.2
scikit-optimize==0.8.1
scipy==1.4.1
scraperapi-sdk==0.2.2
seaborn==0.11.0
SecretStorage==2.3.1
selenium==3.141.0
sentence-transformers==0.2.5
sentencepiece==0.1.91
seqeval==1.2.2
simplejson==3.13.2
six==1.15.0
sklearn==0.0
sortedcontainers==2.3.0
soupsieve==1.7.3
SQLAlchemy==1.3.20
sqlalchemy-mixins==1.2.1
tblib==1.7.0
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorboardcolab==0.0.22
tensorboardX==2.1
tensorflow==1.15.0
tensorflow-addons==0.8.3
tensorflow-datasets==2.1.0
tensorflow-estimator==1.15.1
tensorflow-gcs-config==2.3.0
tensorflow-hub==0.9.0
tensorflow-metadata==0.24.0
tensorflow-privacy==0.2.2
tensorflow-probability==0.9.0
termcolor==1.1.0
terminaltables==3.1.0
threadpoolctl==2.1.0
tifffile==2020.9.3
tokenizers==0.9.4
toml==0.10.2
toolz==0.11.1
torch==1.7.0
torchaudio==0.7.0
torchvision==0.2.2.post3
tornado==6.1
tqdm==4.54.1
transaction==3.0.0
transformers==2.3.0
typed-ast==1.3.1
typeguard==2.12.0
typing==3.6.6
typing-extensions==3.7.4.3

umsgpack==0.1.0

urllib3==1.25.9
vine==1.3.0
virtualenv==15.1.0
w3lib==1.22.0
wadllib==1.3.2
watchtower==1.0.0
webcolors==1.10
webptools==0.0.3
Werkzeug==1.0.1
wrapt==1.12.1
xgboost==1.2.0
yacs==0.1.8
zict==2.0.0
zipp==3.1.0
zope.interface==4.7.1